Clients in the sidebar of the dashboard

+++ SMM/src/app/services/backend-comunication.service.ts
@@ -1,4 +1,4 @@
-import { HttpClient } from '@angular/common/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { environment } from 'src/environments/environment.development';
@@ -13,23 +13,33 @@ export class BackendComunicationService {
 
   /* All the HTTP methods */
 
-  get(endPoint: string): Observable<Object> {
-    return this.http.get(`${this.baseUrl}/${endPoint}`);
+  get(endPoint: string, token?: string): Observable<Object> {
+    if(!token) return this.http.get(`${this.baseUrl}/${endPoint}`);
+    const httpHeaders: HttpHeaders = new HttpHeaders({ Authorization: token });
+    return this.http.get(`${this.baseUrl}/${endPoint}`, { headers: httpHeaders });
   }
 
-  post(endPoint: string, body: Object): Observable<Object> {
-    return this.http.post(`${this.baseUrl}/${endPoint}`, body);
+  post(endPoint: string, body: Object, token?: string): Observable<Object> {
+    if(!token) return this.http.post(`${this.baseUrl}/${endPoint}`, body);
+    const httpHeaders: HttpHeaders = new HttpHeaders({ Authorization: token });
+    return this.http.post(`${this.baseUrl}/${endPoint}`, { headers: httpHeaders });
   }
 
-  put(endPoint: string, body: Object): Observable<Object> {
-    return this.http.put(`${this.baseUrl}/${endPoint}`, body);
+  put(endPoint: string, body: Object, token?: string): Observable<Object> {
+    if(!token) return this.http.put(`${this.baseUrl}/${endPoint}`, body);
+    const httpHeaders: HttpHeaders = new HttpHeaders({ Authorization: token });
+    return this.http.put(`${this.baseUrl}/${endPoint}`, { headers: httpHeaders });
   }
 
-  delete(endPoint: string): Observable<Object> {
-    return this.http.delete(`${this.baseUrl}/${endPoint}`);
+  delete(endPoint: string, token?: string): Observable<Object> {
+    if(!token) return this.http.delete(`${this.baseUrl}/${endPoint}`);
+    const httpHeaders: HttpHeaders = new HttpHeaders({ Authorization: token });
+    return this.http.delete(`${this.baseUrl}/${endPoint}`, { headers: httpHeaders });
   }
 
-  patch(endPoint: string, body: Object): Observable<Object> {
-    return this.http.patch(`${this.baseUrl}/${endPoint}`, body);
+  patch(endPoint: string, body: Object, token?: string): Observable<Object> {
+    if(!token) return this.http.patch(`${this.baseUrl}/${endPoint}`, body);
+    const httpHeaders: HttpHeaders = new HttpHeaders({ Authorization: token });
+    return this.http.patch(`${this.baseUrl}/${endPoint}`, { headers: httpHeaders });
   }
 }
